requirements.txt
# Ingesta y notebooks (Track B). El código del framework vive en agentes/ (sus dependencias allí).
datasets>=2.16.0
huggingface-hub>=0.20.0
tqdm>=4.66.0
pandas>=2.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
